requirements.txt
amqp==2.6.1
aniso8601==7.0.0
asgiref==3.2.10
attrs==20.1.0
Babel==2.8.0
beautifulsoup4==4.9.1
billiard==3.6.3.0
boto3==1.14.53
boto3-stubs==1.14.53.0
botocore==1.17.53
braintree==4.3.0
cachetools==4.1.1
cairocffi==1.1.0
CairoSVG==2.4.2
celery==4.4.7
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
cssselect2==0.3.0
defusedxml==0.6.0
dj-config-url==0.1.1
dj-database-url==0.5.0
dj-email-url==1.0.1
Django==3.1
django-appconf==1.0.4
django-cache-url==3.1.2
django-countries==6.1.3
django-debug-toolbar==2.2
django-debug-toolbar-request-history==0.1.3
django-environ==0.4.5
django-filter==2.3.0
django-graphiql-debug-toolbar==0.1.4
django-js-asset==1.2.2
django-measurement==3.2.3
django-mptt==0.11.0
django-phonenumber-field==5.0.0
django-prices==2.2.0
django-prices-openexchangerates==1.1.0
django-prices-vatlayer==1.1.0
django-redis==4.12.1
django-render-block==0.7
django-templated-email==2.3.0
django-versatileimagefield==2.0
docutils==0.15.2
draftjs-sanitizer==1.0.0
enmerkar==0.7.1
Faker==4.1.2
freezegun==0.3.15
google==3.0.0
google-api-core==1.22.1
google-auth==1.21.0
google-cloud-pubsub==1.7.0
google-i18n-address==2.3.5
googleapis-common-protos==1.52.0
graphene==2.1.8
graphene-django==2.13.0
graphene-federation==0.1.0
graphql-core==2.3.2
graphql-relay==2.0.1
grpc-google-iam-v1==0.12.3
grpcio==1.31.0
html5lib==1.1
idna==2.10
iniconfig==1.0.1
jaeger-client==4.3.0
jmespath==0.10.0
jsonfield==3.1.0
kombu==4.6.11
maxminddb==2.0.2
maxminddb-geolite2==2018.703
measurement==3.2.0
more-itertools==8.4.0
mpmath==1.1.0
mypy-boto3==1.14.53.0
oauthlib==3.1.0
opentracing==2.3.0
packaging==20.4
phonenumberslite==8.12.8
Pillow==7.2.0
pluggy==0.13.1
prices==1.0.0
promise==2.3
protobuf==3.13.0
psycopg2-binary==2.8.5
py==1.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==1.7.1
pyparsing==2.4.7
Pyphen==0.9.5
pytest==6.0.1
python-dateutil==2.8.1
python-json-logger==0.1.11
python-magic==0.4.18
pytimeparse==1.1.8
pytz==2020.1
redis==3.5.3
requests==2.24.0
rsa==4.6
Rx==1.6.1
s3transfer==0.3.3
sentry-sdk==0.17.1
singledispatch==3.4.0.3
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
stripe==2.50.0
sympy==1.6.2
text-unidecode==1.3
threadloop==1.0.2
thrift==0.13.0
tinycss2==1.0.2
toml==0.10.1
tornado==6.0.4
typing==3.7.4.3
ua-parser==0.10.0
Unidecode==1.1.1
urllib3==1.25.10
user-agents==2.2.0
vine==1.3.0
WeasyPrint==51
webencodings==0.5.1
openpyxl==3.0.7
xmltodict==0.12.0
